pipeline/requirements.txt
# Shattered Horizon 2032 — Art Pipeline Dependencies

# CLI framework
typer[all]>=0.9.0

# Terminal output formatting
rich>=13.0.0

# Async HTTP client
httpx>=0.25.0

# Image processing (style checker)
Pillow>=10.0.0
numpy>=1.24.0

# CLIP scoring (optional — style checker degrades gracefully without it)
# open-clip-torch>=2.20.0
# torch>=2.0.0

# Environment variable management
python-dotenv>=1.0.0
